Add PlaceHolderAPI support
