+++ README.md
@@ -1,3 +1,7 @@
 # MidensBounties
 Source code of the Miden's Bounties plugin. Spigot page can be found at https://www.spigotmc.org/resources/midens-bounties-simple-bounty-system.96070/
 Downloads are available Here and on the Spigot page. Experimental versions will be found here and on the discord server (https://discord.gg/xBS7jk6Wtg)
+
+## PlaceHolderAPI
+
+`%midensbounty_bounty%` - Returns the users current bounty (e.g. `$0`)
