+++ src/main/resources/plugin.yml
@@ -1,8 +1,9 @@
 name: MidensBounties
-main: net.cozycosmos.midensbounties.Main
+main: net.cozycosmos.midensbounties.MidensBounties
 author: Miden
 version: 1.1.1
 depend: [Vault]
+softdepend: [PlaceHolderAPI]
 
 commands:
   bounty:
